Conflicts:
	core/implementation/core-api/pom.xml
	core/implementation/core-base/pom.xml
	core/implementation/core-browser/pom.xml
	core/implementation/core-communication/pom.xml
	core/implementation/core-configuration/pom.xml
	core/implementation/core-googlemaps/pom.xml
	core/implementation/core-logging/pom.xml
	core/implementation/core-monitoring/pom.xml
	core/implementation/core-paypal/pom.xml
	core/implementation/core-persistence/pom.xml
	core/implementation/core-reporting/pom.xml
	core/implementation/core-security-oauth/pom.xml
	core/implementation/core-security/pom.xml
	core/implementation/core-testing/pom.xml
	core/implementation/pom.xml
	core/pom.xml
	droid/implementation/droid-api/pom.xml
	droid/implementation/droid-location/pom.xml
	droid/implementation/droid-logging/pom.xml
	droid/implementation/droid-persistence/pom.xml
	droid/implementation/droid-service/pom.xml
	droid/implementation/droid-ui/pom.xml
	droid/implementation/droid-utils/pom.xml
	droid/implementation/pom.xml
	droid/pom.xml
	indy/implementation/indy-base/pom.xml
	indy/implementation/indy-security/pom.xml
	indy/implementation/pom.xml
	indy/pom.xml
	pom.xml

+++ core/implementation/core-security-oauth/src/main/java/sk/qbsw/core/security/oauth/service/impl/ATokenService.java
@@ -119,11 +119,14 @@ abstract class ATokenService
 	{
 		if (authenticationToken != null)
 		{
-			if ( (authenticationToken.getIp() != null && authenticationToken.getIp().equals(ip) == false) || (authenticationToken.getIp() == null && ip != null))
+			if (oauthConfigurator.isIpIgnored() == false)
 			{
-				authenticationTokenDao.remove(authenticationToken);
-				LOGGER.warn("The authentication token {} for user {} and device {} deleted because of invalid ip", authenticationToken.getToken(), authenticationToken.getUser().getLogin(), authenticationToken.getDeviceId());
-				throw new CBusinessException(ECoreErrorResponse.AUTHENTICATION_TOKEN_INVALIDATED);
+				if ( (authenticationToken.getIp() != null && authenticationToken.getIp().equals(ip) == false) || (authenticationToken.getIp() == null && ip != null))
+				{
+					authenticationTokenDao.remove(authenticationToken);
+					LOGGER.warn("The authentication token {} for user {} and device {} deleted because of invalid ip", authenticationToken.getToken(), authenticationToken.getUser().getLogin(), authenticationToken.getDeviceId());
+					throw new CBusinessException(ECoreErrorResponse.AUTHENTICATION_TOKEN_INVALIDATED);
+				}
 			}
 
 			checkAuthenticationTokenValidity(authenticationToken);
